+++ src/main/java/lk/inventory/repository/StockTransactionRepository.java
@@ -3,5 +3,5 @@ package lk.inventory.repository;
 import lk.inventory.entity.StockTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockTransactionRepository extends JpaRepository<StockTransaction,Integer> {
+public interface StockTransactionRepository extends JpaRepository<StockTransaction,Long> {
 }
